create intro game

+++ 04-Collision/PlayScence.cpp
@@ -205,7 +205,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line) // hàm dùng để khởi t
 		{
 			float r = (float)atof(tokens[4].c_str());
 			float b = (float)atof(tokens[5].c_str());
-			float scene_id = (float)atoi(tokens[6].c_str());
+			int scene_id = (int)atoi(tokens[6].c_str());
 			obj = new CPortal(x, y, r, b, scene_id);
 			break;
 		}
