+++ 04-Collision/Portal.cpp
@@ -3,7 +3,7 @@
 #include "Textures.h"
 
 
-CPortal::CPortal(float l, float t, float r, float b, float scene_id)
+CPortal::CPortal(float l, float t, float r, float b, int scene_id)
 {
 	this->scene_id = scene_id;
 	this->x = l;
